1440: Move OB error codes to version agnostic pkg
OBStandardErrorCodes1 and StandardErrorCode have been moved to pkg: uk.org.openbanking.datamodel.error and can be reused across versions.

OBError1 and OBErrorResponse1 for v4 have been moved from the common pkg into: uk.org.openbanking.datamodel.v4.error for consistency with the v3 pkg structure.

https://github.com/SecureApiGateway/SecureApiGateway/issues/1440
