+++ secure-api-gateway-ob-uk-common-obie-datamodel/src/main/java/uk/org/openbanking/datamodel/error/StandardErrorCode.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package uk.org.openbanking.datamodel.v3.error;
+package uk.org.openbanking.datamodel.error;
 
 /**
  * This interface allows you to extend the enum OB standard error code
