--- secure-api-gateway-ob-uk-common-obie-datamodel/src/main/java/uk/org/openbanking/jackson/v3/account/OBSupplementaryData1Serializer.java
@@ -1,43 +1,0 @@
-/*
- * Copyright © 2020-2024 ForgeRock AS (obst@forgerock.com)
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package uk.org.openbanking.jackson.v3.account;
-import com.fasterxml.jackson.core.JsonGenerator;
-import com.fasterxml.jackson.databind.JsonNode;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.fasterxml.jackson.databind.SerializerProvider;
-import com.fasterxml.jackson.databind.ser.std.StdSerializer;
-import uk.org.openbanking.datamodel.v3.common.OBSupplementaryData1;
-
-import java.io.IOException;
-
-public class OBSupplementaryData1Serializer extends StdSerializer<OBSupplementaryData1> {
-
-    public OBSupplementaryData1Serializer() {
-        this(null);
-    }
-
-    public OBSupplementaryData1Serializer(Class<OBSupplementaryData1> t) {
-        super(t);
-    }
-
-    @Override
-    public void serialize(OBSupplementaryData1 supplementaryData1, JsonGenerator jsonGenerator, SerializerProvider serializerProvider)
-            throws IOException {
-        ObjectMapper mapper = new ObjectMapper();
-        JsonNode actualObj = mapper.readTree(supplementaryData1.getData());
-        jsonGenerator.writeObject(actualObj);
-    }
-}
